+++ packages/gatsby/src/utils/commands/jobs-manager.ts
@@ -1,5 +1,6 @@
-import { emitter, store } from "../redux"
-import { waitUntilAllJobsComplete as waitUntilAllJobsV2Complete } from "../utils/jobs-manager"
+import { emitter, store } from "../../redux"
+
+import { waitUntilAllJobsComplete as waitUntilAllJobsV2Complete } from "../jobs-manager"
 
 export const waitUntilAllJobsComplete = (): Promise<void> => {
   const jobsV1Promise = new Promise(resolve => {
